Package rusticl as mesa-opencl-icd
Mesa 22.3 provides a new OpenCL implementation called rusticl;
written in Rust. It is a modern alternative to the "Clover" OpenCL
implementation. To be clear, Rust is only required during the build
process: the final libraries do not need the Rust runtime.

Package rusticl in parallel to the Clover implementation such that
we can ship the rusticl driver to users who explicitly install it
and not impact those users who still use Clover.

According to [1], Clover is due to be removed from Mesa at some
point, so we should package both rusticl and clover in the same
package to help users transitition to rusticl.

[1]: https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/19385
Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>